Add files via upload
We publish the code of the paper entitled "FASPR: A Fast Sparse Phase Retrieval Algorithm via the Epigraph Concept".  In this paper, we proposed a fast sparsity-based phase retrieval algorithm. This algorithm can recover the real-valued image and the complex-valued image from the undersampled and noisy diffraction pattern. Most importantly, few parameters need to be tuned finely in the proposed algorithm.